+++ shaarli_client/__init__.py
@@ -1,5 +1,5 @@
 """shaarli-client"""
 __title__ = 'shaarli_client'
 __brief__ = 'CLI to interact with a Shaarli instance'
-__version__ = '0.1.0'
+__version__ = '0.2.0'
 __author__ = 'The Shaarli Community'