+++ docs/changelog.rst
@@ -9,8 +9,11 @@ The format is based on `Keep a Changelog`_ and this project adheres to
 .. _Keep A Changelog: http://keepachangelog.com/
 .. _Semantic Versioning: http://semver.org/
 
-`v0.2.0 <https://github.com/shaarli/python-shaarli-client/releases/tag/v0.2.0>`_ - UNPUBLISHED
+`v0.3.0 <https://github.com/shaarli/python-shaarli-client/releases/tag/v0.3.0>`_ - UNPUBLISHED
 ----------------------------------------------------------------------------------------------
+
+`v0.2.0 <https://github.com/shaarli/python-shaarli-client/releases/tag/v0.2.0>`_ - 2017-04-09
+---------------------------------------------------------------------------------------------
 
 **Added:**